+++ config/default.php
@@ -686,6 +686,10 @@ $config['cops_epub_reader'] = 'monocle';
  * '' (default)
  * 'comic-reader.html?url='
  * 'codedread-kthoom.html?bookUri=' (with mikespub/codedread-kthoom package)
+ *
+ * Note: for kthoom please install the package with composer:
+ *
+ * $ composer require mikespub/codedread-kthoom
  */
 $config['cops_comic_reader'] = '';